FINANCE/Agentes/requirements.txt
pytesseract
Pillow
openai
python-dotenv
watchdog
pdf2image
